AIML/requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
requests
python-dotenv
pillow
